fix: format decimal
format decimal

+++ lib/ui/cloud_evaluation_GUI.m
@@ -723,7 +723,7 @@ metadata.processor_author = progInfo.Author;
 %% save the results
 cloud_starttime = datenum(handles.cloud_start_tb.String, 'yyyy-mm-dd HH:MM:SS');
 cloud_stoptime = datenum(handles.cloud_stop_tb.String, 'yyyy-mm-dd HH:MM:SS');
-saveFile = fullfile(handles.settings.saveDir, sprintf('cloud_eval_output_%s-%s_%05d-%05d_sm%04d.mat', datestr(cloud_starttime, 'yyyymmdd_HHMM'), datestr(cloud_stoptime, 'yyyymmdd_HHMM'), overviewInfo.CBH * 1000, overviewInfo.CTH * 1000, retrievalInfo.smooth_window));
+saveFile = fullfile(handles.settings.saveDir, sprintf('cloud_eval_output_%s-%s_%05d-%05d_sm%04d.mat', datestr(cloud_starttime, 'yyyymmdd_HHMM'), datestr(cloud_stoptime, 'yyyymmdd_HHMM'), int32(overviewInfo.CBH * 1000), int32(overviewInfo.CTH * 1000), retrievalInfo.smooth_window));
 save(saveFile, 'overviewInfo', 'retrievalInfo', 'settings', 'metadata', 'widgetInfo', '-v6');
 
 %% update status
@@ -742,7 +742,7 @@ function export_fig_btn_Callback(hObject, eventdata, handles)
 
 cloud_starttime = datenum(handles.cloud_start_tb.String, 'yyyy-mm-dd HH:MM:SS');
 cloud_stoptime = datenum(handles.cloud_stop_tb.String, 'yyyy-mm-dd HH:MM:SS');
-saveFile = fullfile(handles.settings.saveDir, sprintf('cloud_eval_output_%s-%s_%05d-%05d_sm%s.png', datestr(cloud_starttime, 'yyyymmdd_HHMM'), datestr(cloud_stoptime, 'yyyymmdd_HHMM'), str2double(handles.cloud_base_tb.String) * 1000, str2double(handles.cloud_top_tb.String) * 1000, handles.smoothwin_tb.String));
+saveFile = fullfile(handles.settings.saveDir, sprintf('cloud_eval_output_%s-%s_%05d-%05d_sm%s.png', datestr(cloud_starttime, 'yyyymmdd_HHMM'), datestr(cloud_stoptime, 'yyyymmdd_HHMM'), int32(str2double(handles.cloud_base_tb.String) * 1000), int32(str2double(handles.cloud_top_tb.String) * 1000), handles.smoothwin_tb.String));
 export_fig(gcf, saveFile, '-r300');
 
 %% update status
